chore: move config, input_data and onnx to input submodule
