+++ rex_xai/mutant.py
@@ -13,7 +13,7 @@ import matplotlib.pyplot as plt
 from rex_xai.box import Box
 from rex_xai.prediction import Prediction
 from rex_xai.utils.logger import logger
-from rex_xai.input_data import Data
+from rex_xai.input.input_data import Data
 from rex_xai.utils._utils import add_boundaries, set_boolean_mask_value
 
 __combinations = [
