+++ rex_xai/responsibility.py
@@ -17,8 +17,8 @@ except ImportError:
 
 
 from rex_xai.box import average_box_size, initialise_tree
-from rex_xai.config import CausalArgs, Queue
-from rex_xai.input_data import Data
+from rex_xai.input.config import CausalArgs, Queue
+from rex_xai.input.input_data import Data
 from rex_xai.utils.logger import logger
 from rex_xai.mutant import Mutant, get_combinations, _apply_to_data
 from rex_xai.resp_maps import ResponsibilityMaps
