+++ scripts/spectral_pytorch.py
@@ -6,7 +6,7 @@ import torch as tt
 import torch.nn as nn
 import torch.nn.functional as F
 
-from rex_xai.input_data import Data
+from rex_xai.input.input_data import Data
 from rex_xai.prediction import from_pytorch_tensor
 
 class ConvNet(nn.Module):
