+++ tests/scripts/pytorch_3d.py
@@ -2,7 +2,7 @@
 from onnxruntime.tools.ort_format_model.ort_flatbuffers_py.fbs.Tensor import Tensor
 import torch as tt
 import torch.nn.functional as F
-from rex_xai.input_data import Data
+from rex_xai.input.input_data import Data
 from rex_xai.prediction import from_pytorch_tensor
 from monai.networks.nets import DenseNet121
 from monai.transforms import Resize, LoadImage
