+++ tests/unit_tests/config_test.py
@@ -2,7 +2,7 @@ import copy
 
 import pytest
 from rex_xai.utils._utils import Queue, Strategy
-from rex_xai.config import CausalArgs, process_config_dict, read_config_file
+from rex_xai.input.config import CausalArgs, process_config_dict, read_config_file
 from rex_xai.distributions import Distribution
