+++ tests/unit_tests/visualisation_test.py
@@ -3,7 +3,7 @@ import os
 import torch as tt
 from rex_xai.visualisation import save_image, voxel_plot
 
-from rex_xai.config import CausalArgs
+from rex_xai.input.config import CausalArgs
 
 
 def test_surface(exp_extracted, tmp_path):
